requirements/base.txt
appdirs==1.4.4
attrs==20.3.0
autopep8==1.5.7
Babel==2.6.0
beautifulsoup4==4.9.3
black==21.6b0
cached-property==1.5.2
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
decorator==4.3.0
defusedxml==0.7.1
docutils==0.14
ebaysdk==2.1.5
feedparser==5.2.1
flake8==3.9.2
fpdf==1.7.2
future==0.18.2
gevent==1.5.0
greenlet==0.4.15
html2text==2018.1.9
idna==2.8
img2pdf==0.4.0
isodate==0.6.0
Jinja2==2.10.1
libsass==0.17.0
lxml==4.6.3
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
mypy-extensions==0.4.3
num2words==0.5.6
ofxparse==0.19
passlib==1.7.1
pathspec==0.8.1
pbr==5.6.0
pdf417gen==0.7.1
phonenumbers==8.12.22
pikepdf==2.11.4
Pillow==8.2.0
#pkg-resources==0.0.0
polib==1.1.0
psutil==5.6.6
psycopg2==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydot==1.4.1
pyflakes==2.3.1
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==3.4
PySimpleSOAP==1.16.2
python-dateutil==2.7.3
python-ldap==3.1.0
python-stdnum==1.16
pytz==2019.1
pyusb==1.0.2
qrcode==6.1
regex==2021.4.4
reportlab==3.5.13
requests==2.21.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.6.1
Werkzeug==0.14.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmlsec==1.3.10
zeep==3.2.0
